requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
distro==1.4.0
fabric==2.5.0
flake8==3.8.3
hiredis==1.0.1
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.1
invoke==1.4.1
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.5.1
mccabe==0.6.1
msgpack==0.6.2
multidict==4.7.6
packaging==20.3
paramiko==2.7.1
parsel==1.6.0
pep517==0.8.2
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.6
python-dotenv==0.13.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
urllib3==1.25.9
w3lib==1.22.0
webencodings==0.5.1
yarl==1.4.2
zipp==3.1.0
